requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
py-radix==0.10.0
requests==2.31.0
requests-futures==1.0.1
urllib3==2.1.0
